% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{openPoland}
\alias{openPoland}
\alias{openPoland-package}
\title{Generate R documentation from inline comments.}
\description{
An R package for communication with OpenPoland.net API
}
\seealso{
\itemize{
  \item \url{https://github.com/kalimu/openPoland}
  \item \url{https://openpoland.net/welcome/}
}
}

