% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandAPI}
\alias{openPolandAPI}
\title{A tool function that handles communication with API over multiple pages}
\usage{
openPolandAPI(url, token = NULL, query = NULL, nts = NULL, year = NULL)
}
\description{
\code{openPolandAPI} modifies url string and binds data from multiple pages.
}
\keyword{internal}

