% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandData}
\alias{openPolandData}
\title{Search openPoland.net database for a specific dataset.}
\usage{
openPolandData(id, token, nts = NULL, year = NULL)
}
\description{
\code{openPolandSearch} returns a data frame with information of available datasets that were found with a given quary.
}

