% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandFilter}
\alias{openPolandFilter}
\title{Filter a dataset from openPoland API.}
\usage{
openPolandFilter(data = NULL, nts = NULL, unit = NULL, year = NULL,
  name = NULL)
}
\arguments{
\item{data}{A data table that is a result of \code{\link{openPolandData}} function.}

\item{nts}{A character string.}

\item{unit}{A character string of predefined values: "region", "voivodship", "subregion", "powiat", "gmina".}

\item{year}{A character string.}

\item{name}{A character string.}
}
\value{
A data table object. The first column is a NTS id of territorial unit. The second column is a common name of territorial unit. Then there are from 1 to 5 columns with dimensions labels. All dimensions in a given dataset can be previewed by the \code{\link{openPolandMeta}} function. Last four columns of the data table are: year, measure unit, value and data attribute.
}
\description{
\code{openPolandFilter} filter a dataset with given criterias and returns a data frame.
}
\examples{
\dontrun{

# token for API authorization
token = "[alpha-numeric-string-you-get-after-registration-on-https://openpoland.net]"

openPolandMeta(id =  1944, token = token)

# a whole, unfiltered dataset
data = openPolandData(id = 1944, token = token)

unique(openPolandFilter(data = data, unit="region")$name)
unique(openPolandFilter(data = data, unit="voivodship")$name)
unique(openPolandFilter(data = data, unit="subregion")$name)
unique(openPolandFilter(data = data, unit="powiat")$name)
unique(openPolandFilter(data = data, unit="gmina")$name)

unique(openPolandFilter(data = data, nts=5)$name)

openPolandFilter(data = data, name="Warszawa")

}
}
\seealso{
\code{\link{openPolandData}} for downloading a selected dataset via openPoland API, \code{\link{openPolandMeta}} for getting meta information about datasets, and \code{\link{openPolandSearch}} for searching database with a given query.
}

