% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandMeta}
\alias{openPolandMeta}
\title{Search openPoland.net database for a specific dataset.}
\usage{
openPolandMeta(id, token, verbose = TRUE)
}
\description{
\code{openPolandSearch} returns a data frame with information of available datasets that were found with a given quary.
}

