% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandQuery}
\alias{openPolandQuery}
\title{A tool function that sends a single query to the OpenPoland API.}
\usage{
openPolandQuery(url, token = NULL, meta = FALSE, query = NULL,
  nts = NULL, year = NULL)
}
\description{
\code{openPolandQuery} handles a different type of content in API responses.
}
\keyword{internal}

