% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandSearch}
\alias{openPolandSearch}
\title{Search openPoland.net database for a specific dataset.}
\usage{
openPolandSearch(query = NULL, token = NULL, subQuery = NULL)
}
\arguments{
\item{query}{A character string.}

\item{token}{A character string.}

\item{subQuery}{A characters string.}
}
\value{
A data table object with datasets ids and titles that have a query string inside them.
}
\description{
\code{openPolandSearch} returns a data frame with information of available datasets that were found with a given quary.
}
\details{
The query from \code{openPolandSearch} function is not case-sensitive.
The token authorization is non-obligatory for database searching. If you want to use Polish diacritis in the query remeber to check if they are properly encoded ("UTF-8") before they are sent to API. If you use Polish localization the function \code{\link{enc2utf8}} should help.
The query parameter is send to API, but the subQuery parameter is search locally after downloading the dataset lists from API.
}
\examples{
\dontrun{
openPolandSearch() # returns API demo list of datasets

openPolandSearch(query = 'telef')

openPolandSearch(query = "nauczy", subQuery = "akad")

# Polish diacritics are also available
openPolandSearch(query = enc2utf8('ludno\\u015B\\u0107'))
}
}

