% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandData}
\alias{openPolandData}
\title{Loads a full or filtered data from a selected dataset via openPoland API.}
\usage{
openPolandData(id = NULL, token = NULL, nts = NULL, year = NULL,
  dimLabels = TRUE, remove_duplicated = TRUE)
}
\arguments{
\item{id}{A character string.}

\item{token}{A characters string.}

\item{nts}{A character string.}

\item{year}{A character string.}

\item{dimLabels}{Logical value. if TRUE (default) the function automatically merges meta data (dimensions names and labels) with raw data.}

\item{remove_duplicated}{A logical value. If TRUE (default) removes all duplicated records from the datasets.}
}
\value{
A data table object. The first column is a NTS id of territorial unit. The second column is a common name of territorial unit. Then there are from 1 to 5 columns with dimensions labels. All dimensions in a given dataset can be previewed by the \code{\link{openPolandMeta}} function. Last four columns of the data table are: year, measure unit, value and data attribute.
}
\description{
\code{openPolandData} returns a data frame with data from a dataset with given ID. The data could be filtered by year or NTS id.
}
\examples{
\dontrun{

openPolandSearch("akad")
#     id                                     title
# 1 1836 Nauczyciele akademiccy (NTS-2, 1999-2012)

# token for API authorization
token = "[alpha-numeric-string-you-get-after-registration-on-https://openpoland.net]"

# a fltered dataset with data for 1999 year in Central Region (NTS = 1000000000)
openPolandData(id = 1836, token = token,nts = '1000000000', year='1999')

# a whole, unfiltered dataset
openPolandData(id = 1836, token = token)
}
}
\seealso{
\code{\link{openPolandMeta}} for getting meta information about datasets, \code{\link{openPolandSearch}} for searching database with a given query and \code{\link{openPolandFilter}} for filtering dataset with given criterias.
}

