% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandFilter}
\alias{openPolandFilter}
\title{Filter a dataset from openPoland API.}
\usage{
openPolandFilter(data = NULL, nts = NULL, unit = NULL, year = NULL,
  name = NULL)
}
\arguments{
\item{data}{A data table that is a result of \code{\link{openPolandData}} function.}

\item{nts}{A character string.}

\item{unit}{A character string of predefined values: "region", "voivodship", "subregion", "powiat", "gmina".}

\item{year}{A character string.}

\item{name}{A character string.}
}
\value{
A data table object. The first column is a NTS id of territorial unit. The second column is a common name of territorial unit. Then there are from 1 to 5 columns with dimensions labels. All dimensions in a given dataset can be previewed by the \code{\link{openPolandMeta}} function. Last four columns of the data table are: year, measure unit, value and data attribute.
}
\description{
\code{openPolandFilter} filter a dataset with given criterias and returns a data frame.
}
\examples{
\dontrun{

openPolandSearch("akad")
#     id                                     title
# 1 1836 Nauczyciele akademiccy (NTS-2, 1999-2012)

# token for API authorization
token = "[alpha-numeric-string-you-get-after-registration-on-https://openpoland.net]"

# a fltered dataset with data for 1999 year in Central Region (NTS = 1000000000)
openPolandData(id = 1836, token = token,nts = '1000000000', year='1999')

# a whole, unfiltered dataset
openPolandData(id = 1836, token = token)
}
}
\seealso{
\code{\link{openPolandData}} for downloading a selected dataset via openPoland API, \code{\link{openPolandMeta}} for getting meta information about datasets, and \code{\link{openPolandSearch}} for searching database with a given query.
}

