% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandMeta}
\alias{openPolandMeta}
\title{Shows meta information about a specific dataset.}
\usage{
openPolandMeta(id = NULL, token = NULL, verbose = TRUE)
}
\arguments{
\item{id}{A character string.}

\item{token}{A characters string.}

\item{verbose}{If TRUE prints meta info on the console and returns nothing. If FALSE doesn't print meta but returns it as a list object.}
}
\value{
If verbose = FALSE returns a list object with meta information about a selected dataset.
}
\description{
\code{openPolandMeta} returns a list with meta information of a specific dataset and prints it on the console in a readable form.
}
\examples{
\dontrun{

openPolandSearch("akad")
#     id                                     title
# 1 1836 Nauczyciele akademiccy (NTS-2, 1999-2012)

# token for API authorization
token = "[alpha-numeric-string-you-get-after-registration-on-https://openpoland.net]"

# print meta data on the console
openPolandMeta(id = 1836, token = token)

# don't print and return only list object
openPolandMeta(id = 1836, token = token, verbose = FALSE)

}
}
\seealso{
\code{\link{openPolandSearch}} for  for searching database with a given query and \code{\link{openPolandData}} for downloading a selected dataset via openPoland API.
}

