% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openPolandSearch}
\alias{openPolandSearch}
\title{Search openPoland.net database for a specific dataset.}
\usage{
openPolandSearch(query = NULL, token = NULL, subQuery = NULL)
}
\arguments{
\item{query}{A character string.}

\item{token}{A character string.}

\item{subQuery}{A characters string.}
}
\value{
A data table object with datasets ids and titles that have a query string inside them.
}
\description{
\code{openPolandSearch} returns a data frame with information of available datasets that were found with a given quary.
}
\details{
The query from \code{openPolandSearch} function is not case-sensitive.
The token authorization is non-obligatory for database searching.
Polish diacritics should also worked within a query.
The query parameter is send to API, but the subQuery parameter is search locally after downloading the dataset lists from API.
}
\examples{
\dontrun{
openPolandSearch() # returns list of all datasets

openPolandSearch(query = 'telef')

openPolandSearch(query = "nauczy", subQuery = "akad")

# Polish diacritics are also available but remember
# to set Polish localization first if you want to work with properly
# formated Polish titles of datasets.
Sys.setlocale("LC_ALL", "Polish")
openPolandSearch(query = 'ludno\\\\u015B\\\\u0107')
}
}
\seealso{
\code{\link{openPolandMeta}} for meta data about datasets, \code{\link{openPolandData}} for downloading a selected dataset via openPoland API, and \code{\link{openPolandFilter}} for filtering locally downloaded dataset with given criterias.
}

